Reduce screening common contract surface
MCP screening common carried cache, diagnostics, and response-shaping concerns in the same module as request normalization and filters. This extracts cache and response helpers while preserving common.py re-export compatibility for existing MCP callsites and legacy analysis_screen_core imports.\n\nConstraint: MCP public screening response keys, warnings, and import compatibility must not change.\nRejected: Rewrite screening/us.py, kr.py, and crypto.py imports in the same change | re-export compatibility gives a smaller, safer diff.\nConfidence: high\nScope-risk: moderate\nReversibility: clean\nDirective: Keep common.py as the compatibility boundary until all screening callsites are intentionally migrated.\nTested: uv run --group dev ruff check app/mcp_server/tooling/screening/common.py app/mcp_server/tooling/screening/market_cap_cache.py app/mcp_server/tooling/screening/response.py; uv run --group test pytest tests/test_screening_common.py tests/test_screening_filter_extensions.py tests/test_invest_screener_week_change_filter.py tests/test_screening_consecutive_up_days.py -q (50 passed)\nNot-tested: Full MCP test matrix
